Error Handling for failed to parse json, add_single file functionality fixed with correct instance

+++ update.py
@@ -31,6 +31,7 @@ to-do:
 import os
 import time
 import json
+import simplejson
 import argparse
 import sys
 import shutil
@@ -63,26 +64,37 @@ class SolrInterface(object):
         return raw_query.replace("'", "\'")
 
     def delete_query(self, query, no_confirm=False):
-        global SOLR_INSTANCE
+        # global SOLR_INSTANCE
         # execute delete query
         s = self.solr.search(self.escape_query(query), **{"rows": "0"})
         if s.hits == 0:
             print("No records found in " + SOLR_INSTANCE + " instance. Nothing to delete.  Exiting...")
         else:
-            inMessage = "Are you sure you want to delete {num_recs} record(s) from " + SOLR_INSTANCE + " instance? Y/N: "
+            inMessage = "Are you sure you want to delete {num_recs} record(s) from " + SOLR_INSTANCE + " instance? (Y/N): "
             are_you_sure = input(inMessage.format(num_recs=s.hits))
             if are_you_sure.lower() == "y":
                 self.solr.delete(q=self.escape_query(query))
-                confirmMessage = "Done deleting {num_recs} record(s) from " + SOLR_INSTANCE + " instance..."
+                confirmMessage = "{num_recs} record(s) deleted from " + SOLR_INSTANCE + " instance."
                 print(confirmMessage.format(num_recs=s.hits))
             else:
                 print("Okay, nothing deleted from " + SOLR_INSTANCE + " instance. Exiting...")
+                
 
     def json_to_dict(self, json_doc):
         # read data from one json file
         #print("Reading input file " + json_doc)
-        j = json.load(open(json_doc, "rt", encoding="utf8"))      
-        return j
+        try:
+            j = json.load(open(json_doc, "rt", encoding="utf8"))
+            return j
+        except json.decoder.JSONDecodeError as err:
+            json_name = os.path.basename(json_doc)
+            print(" - {} has a syntax error at line {} column {}. Fix manually and reingest ".format(json_name,err.lineno, err.colno))
+            exit()
+        try:
+            return j
+        except UnboundLocalError:
+            print("Fatal Syntax Error may effect the quality of this ingest")
+        
 
     def add_dict_list_to_solr(self, list_of_dicts):
         # ingest dictionary to Solr
@@ -98,11 +110,12 @@ class SolrInterface(object):
     
 class Update(object):
 
-    defaultPath = "R:\ARCHIVE_PROJECT\SCO_TEST\FinishedOpenMetadata\Eli_ReadyForIngest"
 
     def __init__(self, SOLR_USERNAME, SOLR_PASSWORD, FORCE, SCAN, COLLECTION, PROVENANCE, UUID, INSTANCE, TO_JSON=False, RECURSIVE=False, PURGE=False):
+        global SOLR_INSTANCE
         self.RECURSIVE = RECURSIVE
         self.PURGE = PURGE
+        SOLR_INSTANCE = INSTANCE
         if INSTANCE=="prod":
             SOLR_URL = SOLR_URL_PRODUCTION
         elif INSTANCE=="test":
@@ -111,7 +124,10 @@ class Update(object):
             SOLR_URL = SOLR_URL_DEV
         else: 
             # if instance not specified, default to dev
-            SOLR_URL = SOLR_URL_DEV
+            print("Instance argument is null. Defaulting to dev...")
+            print("Currently this is actually defaulting to test")
+            SOLR_INSTANCE = "test"
+            SOLR_URL = SOLR_URL_TEST
 
         self.solr = SolrInterface(url=SOLR_URL)
         self.uuid = UUID
@@ -120,30 +136,64 @@ class Update(object):
            
     def scan_dict_records(self, list_of_dicts):
         # perform QA checks on dictionary object before it is ingested
-        status = True  
+        status = True
         d = list_of_dicts
-        # Check if required elements have valid data       
-        if(d["dc_identifier_s"] == ""):
-            self.scanCatch.append(os.path.basename(self.currentFile))
-        if(d["layer_slug_s"] == ""):
-            self.scanCatch.append(os.path.basename(self.currentFile))
-        if(d["dc_title_s"] == ""):
-            self.scanCatch.append(os.path.basename(self.currentFile))
-        if("NaN" in d["solr_geom"]):
-            self.scanCatch.append(os.path.basename(self.currentFile))
-        if(d["dct_provenance_s"] == ""):
-            self.scanCatch.append(os.path.basename(self.currentFile))
-        if(d["dc_rights_s"] == ""):
-            self.scanCatch.append(os.path.basename(self.currentFile))
-        if(d["geoblacklight_version"] == ""):
-            self.scanCatch.append(os.path.basename(self.currentFile))
-       # if(d["solr_year_i"] == 9999):
-        #   scanCatch += "solr_year_i\n"
-
+        # Check if required elements have valid data
+        try:
+            if(d["dc_identifier_s"] == ""):
+                #self.totalErrorNumbers = self.totalErrorNumbers + 1
+                self.scanCatch.append(os.path.basename(self.currentFile))
+                self.fileProbList.append("File '" + os.path.basename(self.currentFile)+ "' has null field 'dc_identifier_s'")
+            if(d["layer_slug_s"] == ""):
+                #self.totalErrorNumbers = self.totalErrorNumbers + 1
+                self.scanCatch.append(os.path.basename(self.currentFile))
+                self.fileProbList.append("File '" + os.path.basename(self.currentFile)+ "' has null field 'layer_slug_s'")
+            if(d["layer_slug_s"] == ""):
+                #self.totalErrorNumbers = self.totalErrorNumbers + 1
+                self.scanCatch.append(os.path.basename(self.currentFile))
+                self.fileProbList.append("File '" + os.path.basename(self.currentFile)+ "' has null field 'layer_slug_s'")
+            if("NaN" in d["solr_geom"]):
+                #self.totalErrorNumbers = self.totalErrorNumbers + 1
+                self.scanCatch.append(os.path.basename(self.currentFile))
+                self.fileProbList.append("File '" + os.path.basename(self.currentFile)+ "' has null field 'solr_geom'")
+            if(d["dct_provenance_s"] == ""):
+                #self.totalErrorNumbers = self.totalErrorNumbers + 1
+                self.scanCatch.append(os.path.basename(self.currentFile))
+                self.fileProbList.append("File '" + os.path.basename(self.currentFile)+ "' has null field 'dct_provenance_s'")
+            if(d["dc_rights_s"] == ""):
+                #self.totalErrorNumbers = self.totalErrorNumbers + 1
+                self.scanCatch.append(os.path.basename(self.currentFile))
+                self.fileProbList.append("File '" + os.path.basename(self.currentFile)+ "' has null field 'dc_rights_s'")
+            if(d["geoblacklight_version"] == ""):
+                #self.totalErrorNumbers = self.totalErrorNumbers + 1
+                self.scanCatch.append(os.path.basename(self.currentFile))
+                self.fileProbList.append("File '" + os.path.basename(self.currentFile)+ "' has null field 'geoblacklight_version'")
+            if(d["dc_creator_sm"] == ""):
+                #self.totalErrorNumbers = self.totalErrorNumbers + 1
+                self.scanCatch.append(os.path.basename(self.currentFile))
+                self.fileProbList.append("File '" + os.path.basename(self.currentFile)+ "' has null field 'dc_creator_sm'")
+            if(d["dc_description_s"] == ""):
+                #self.totalErrorNumbers = self.totalErrorNumbers + 1
+                self.scanCatch.append(os.path.basename(self.currentFile))
+                self.fileProbList.append("File '" + os.path.basename(self.currentFile)+ "' has null field 'dc_description_s'")
+            if(d["dc_subject_sm"] == ""):
+                #self.totalErrorNumbers = self.totalErrorNumbers + 1
+                self.scanCatch.append(os.path.basename(self.currentFile))
+                self.fileProbList.append("File '" + os.path.basename(self.currentFile)+ "' has null field 'dc_subject_sm'")
+            if(d["solr_year_i"] == ""):
+                #self.totalErrorNumbers = self.totalErrorNumbers + 1
+                self.scanCatch.append(os.path.basename(self.currentFile))
+                self.fileProbList.append("File '" + os.path.basename(self.currentFile)+ "' has null field 'solr_year_i'")
+               # if(d["solr_year_i"] == 9999):
+                #   scanCatch += "solr_year_i\n"
+        except TypeError:
+            print("QA Scan Error")
+            #exit()
+        
         # If issues, print message and layer ids, set status to false
         if (len(self.scanCatch) != 0):
             status = False
-            #print("\nQA health check failed for " + d["dc_title_s"] +
+            #print("\nQA health check failed for " + d["dc_title_s"])
                   #"\nThe following fields are either missing data or have invalid entries: " + scanCatch) 
             # also send these result to a log file
         
@@ -162,33 +212,36 @@ class Update(object):
             files = glob(os.path.join(start_path, criteria))
         return files
     
-
-    def add(self, path_to_json):
+    def add_single(self, path_to_json):
         self.ingested = False
-        #print("Add Single File Functionality still in Development")
-        global SOLR_INSTANCE
         arg = (path_to_json.strip("\ "))
         if os.path.isfile(arg) and ".json" in arg:
             temp = "temp"
             if not os.path.exists(temp):
+                #print("creating directory temp...")
                 os.mkdir(temp)
+            # get name of file to be ingested
             file=os.path.basename(arg)
             newPath = temp+"/"+file
+            # copy file to temporary folder 'temp'
             shutil.copyfile(arg, newPath)
-            self.force_add(temp)
-            if self.success:
+            # run add_folder function to ingest file in folder 'temp'
+            self.add_folder(temp)
+
+            # delete file from original location so it's copy can be moved back after ingest
+            if self.success and len(os.listdir(temp)) == 0:
                 os.remove(arg)
-                print("No Files to Ingest")
-            if self.ingested == True:
-                #edit this to change where a file goes after it is successfully ingested
+                
+            # move file back if necesary   
+            if self.sortYN.upper() == "Y" and len(os.listdir(temp)) != 0:
                 os.rename(newPath,arg)
+            # delete temporary folder
             shutil.rmtree(temp)   
         else:
-            print("Argument is not a json. Please enter a valid file name.")
+            print("File not found. Please enter a valid path to json")
             
                    
-    def force_add(self, path_to_json):
-        global SOLR_INSTANCE
+    def add_folder(self, path_to_json):
         self.scan(path_to_json)
         files = self.get_files_from_path(path_to_json, criteria="*.json")
         if files and self.success == True:
@@ -196,24 +249,30 @@ class Update(object):
             for i in files:
                 dictAppend = self.solr.json_to_dict(i)
                 self.dicts.append(dictAppend)
-            ingest_result = self.solr.add_dict_list_to_solr(self.dicts)
-            print("%i record(s) successfully ingested into Solr" % (len(self.dicts)))
-            self.ingested = True
-        elif self.success == False:
-            print("Ingest Failed, Ingest Halted")
-        elif not files:
-            print("Ingest Failed, No Files to Ingest")
-        else:
-            print("Ingest Failed, Check JSON Location and Try Again") 
+            if self.readyToIngest == True:
+                confirm = input("Ingest file(s) into %s instance of Solr? (Y/N) " % SOLR_INSTANCE)
+                if confirm.upper() == "Y":
+                    ingest_result = self.solr.add_dict_list_to_solr(self.dicts)
+                    print("{} record(s) successfully ingested into {} instance of Solr".format(len(self.dicts), SOLR_INSTANCE))
+                    self.ingested = True
+                else:
+                    print("Ingest terminated manually. Exiting...")
 
     def scan(self, path_to_json):
+        global SOLR_INSTANCE
+        if SOLR_INSTANCE == "":
+            SOLR_INSTANCE = args.instance
+            print("Solr Instance: %s" % SOLR_INSTANCE)
         self.success = False
         self.scanCatch = []
+        self.fileProbList = []
+        self.sortYN = ""
+        #self.totalErrorNumbers = 0
         # cycle through json files, add them to a dictionary object
         files = self.get_files_from_path(path_to_json, criteria="*.json")
         if files:
             cwd = os.getcwd()
-            path = cwd + "\ for_review"
+            path = cwd + "/for_review"
             self.dicts = []
             print("Performing QA scan...") 
             for i in files:
@@ -221,9 +280,18 @@ class Update(object):
                 dictAppend = self.solr.json_to_dict(i)
                 self.dicts.append(dictAppend)
                 scanHold = self.scan_dict_records(dictAppend)
-            print("QA health check failed on %i file(s): " % (len(self.scanCatch)))
-            sortYN = input("Sort failed files into review folder and continue? (Y/N) ")
-            if (scanHold == False and sortYN.upper()== "Y"):
+            if len(self.scanCatch) != 0:
+                print("QA health check failed on {} file(s) ".format((len(self.scanCatch))))
+                #print("QA health check failed on {} file(s) with {} total errors ".format((len(self.scanCatch)),self.totalErrorNumbers))
+                print("-"*60)
+                for bad_file in self.fileProbList:
+                    print(" - "+ bad_file)
+                print("-"*60)
+                self.sortYN = input("Sort failed files into review folder and continue? (Y/N) ")
+            else:
+                print("QA Health Check Passed!")
+                self.readyToIngest = True
+            if (scanHold == False and self.sortYN.upper()== "Y"):
                     try:
                         if os.path.exists(path):
                             print("Moving files to directory 'for_review'...")
@@ -234,12 +302,12 @@ class Update(object):
                         for file in self.scanCatch:
                             os.rename(path_to_json+ '/' +file, path + '/' + os.path.basename(file))
                         self.success = True
+                        self.readyToIngest = True
                         print( "%i record(s) successfully moved to 'for_review'" % (len(self.scanCatch)))
                     except OSError:
-                        print("holup")
                         print("OS Failure")
-            elif sortYN.upper() == "N":
-                print("No Selected, Ingest Halted")
+            elif self.sortYN.upper() == "N":
+                print("Ingest terminated manually.  Exiting...")
                 self.success = False
             elif len(self.scanCatch) == 0:
                 self.success = True
@@ -270,7 +338,6 @@ class Update(object):
         self.solr.delete_query("*:*")
 
 def main():
-    defaultPath = "test-scenarios\eli-testdata"
     parser = argparse.ArgumentParser()
     parser.add_argument(
         "-r",
@@ -279,11 +346,11 @@ def main():
         help="Recurse into subfolders when adding JSON files.")
     group = parser.add_mutually_exclusive_group(required=True)
     group.add_argument(
-        "-afi",
-        "--addFile",
+        "-as",
+        "--addSingleFile",
         help="Indicate path to a GeoBlacklight JSON file")
     group.add_argument(
-        "-afo",
+        "-a",
         "--addFolder",
         help="Indicate path to folder with GeoBlacklight \
               #JSON files that will be uploaded.")
@@ -315,21 +382,18 @@ def main():
         "-i",
         "--instance",
         choices={"prod", "test","dev"},
-        help="Identify which instance of Solr to use.")           
+        help="Identify which instance of Solr to use.")
     
     args = parser.parse_args()
     interface = Update(SOLR_USERNAME, SOLR_PASSWORD, FORCE=args.addFolder, SCAN=args.scan, COLLECTION=args.delete_collection, PROVENANCE=args.delete_provenance, UUID=args.delete,INSTANCE=args.instance, RECURSIVE=args.recursive, PURGE=args.purge)
     
-    inPath = args.addFile
+    inPath = args.addSingleFile
     if (inPath is not None and inPath[-1] != '\\'):
-            inPath += '\\'
-    global SOLR_INSTANCE
-    SOLR_INSTANCE = args.instance
-    
-    if args.addFile:
-        interface.add(inPath)
+            inPath += '\\'     
+    if args.addSingleFile:
+        interface.add_single(inPath)
     elif args.addFolder:
-        interface.force_add(args.addFolder)
+        interface.add_folder(args.addFolder)
     elif args.scan:
         interface.scan(args.scan)
     elif args.delete_collection:
@@ -341,7 +405,7 @@ def main():
     elif args.purge:
         interface.purge()
     else:
-        print("nothings happened")
+        print("Nothings happened. Exiting...")
         sys.exit(parser.print_help())
 
 if __name__ == "__main__":
